adds Stream Server View, AudienceLayout, simplified invitation logic, RaisedHandsWindow
A Stream of the event can now be created using a special shareToken which skips the name and device setup. Tested and working with OBS's browser feature.

A member of the audience can watch the stream of the event, can chat with other audience members and moderators / participants. The audience member can raise his hand to be invited into the conference by a moderator. This works but does not look optimal

Caution:
I had to remove the "Browser not supported" warning so that OBS can join the session. Everything seems to work fine, but this should stay in our minds.